+++ Changes.md
@@ -6,6 +6,11 @@ Fixes
 
 - Context : Fixed Python `Canceller` lifetime management bug.
 
+API
+---
+
+- Context : Added copy constructor that allows an existing canceller to be omitted.
+
 0.57.5.0 (relative to 0.57.4.1)
 ========